[ADD] Expired exam show label & disable copy link
